Implementing disambiguation experiments

+++ training_ptr_gen/train_util.py
@@ -6,7 +6,7 @@ from data_util import config
 def get_input_from_batch(batch, use_cuda):
   batch_size = len(batch.enc_lens)
 
-  enc_batch = Variable(torch.from_numpy(batch.enc_batch).long())
+  enc_batch = Variable(torch.from_numpy(batch.enc_batch)).float() # remove .long()
   enc_padding_mask = Variable(torch.from_numpy(batch.enc_padding_mask)).float()
   enc_lens = batch.enc_lens
   extra_zeros = None
